Rules Rewrite (Closes #25)
To avoid having the Rule object violate Single Responsibility
Filtering and Annotation were separated.

The actual filter action is now in the "Filters" namespace and
the rules act as simple annotation classes only parsing configuration.

Rewrote tests and added a FilterLoader in the mix to help
find the proper class to handle a rule. This also makes
it possible create new rules that use 'services' to apply themselves.